+++ R/namespace.R
@@ -1,0 +1,3 @@
+#' @import checkr
+#' @importFrom stats lm predict
+NULL