+++ R/package.R
@@ -1,10 +1,15 @@
 #' Kootenay Lake Data
 #'
-#' Annual Rainbow Trout and Kokanee datasets for Kootenay Lake.
+#' Annual Rainbow Trout, Bull Trout and Kokanee datasets for Kootenay Lake.
+#'
+#' It includes the `kl_estimate_na()` function to estimate missing variable values
+#' from a predictor using a linear model.
+#' By default it estimates missing Bull Trout redd counts for Kaslo Creek
+#' using the resistivity counter counts.
 #'
 #' @docType package
 #' @name kootlake
-#' @seealso \code{\link{fishery}}, \code{\link{gerrard}} and \code{\link{kokanee}}
+#' @seealso \code{\link{bulltrout}}, \code{\link{fishery}}, \code{\link{gerrard}} and \code{\link{kokanee}}
 #' @examples
 #' library(ggplot2)
 #'