+++ data-raw/data-raw.R
@@ -1,3 +1,4 @@
 source("data-raw/fishery.R")
 source("data-raw/gerrard.R")
 source("data-raw/kokanee.R")
+source("data-raw/bulltrout.R")