+++ data-raw/fishery.R
@@ -1,9 +1,10 @@
 library(dplyr)
 library(magrittr)
+library(readr)
 library(devtools)
 
 rm(list = ls())
 
-fishery <- read.csv("data-raw/fishery.csv")
+fishery <- read_csv("data-raw/fishery.csv")
 fishery %<>% arrange(Year) %>% select(Year, AnglerHours)
 devtools::use_data(fishery, overwrite = TRUE)