+++ data-raw/gerrard.R
@@ -1,9 +1,10 @@
 library(dplyr)
 library(magrittr)
+library(readr)
 library(devtools)
 
 rm(list = ls())
 
-gerrard <- read.csv("data-raw/gerrard.csv")
+gerrard <- read_csv("data-raw/gerrard.csv")
 gerrard %<>% arrange(Year) %>% select(Year, PeakCount, FishDays)
 devtools::use_data(gerrard, overwrite = TRUE)